chore: move downlevelWorkspace to different file
